pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vectorgov"
version = "0.1.0"
description = "SDK Python para acessar bases de conhecimento jurídico VectorGov"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "VectorGov", email = "contato@vectorgov.io"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["vectorgov", "rag", "retrieval", "legal", "ai", "llm"]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.18"]
google = ["google-generativeai>=0.3"]

[project.urls]
Homepage = "https://vectorgov.io"
Documentation = "https://docs.vectorgov.io"
Repository = "https://github.com/vectorgov/vectorgov-python"
Issues = "https://github.com/vectorgov/vectorgov-python/issues"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*", "examples*", "docs*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=vectorgov --cov-report=term-missing"
